requirements.txt
Django==5.0
Pillow
django-jazzmin
requests